+++ src/assets/src/configurator/BlockType.js
@@ -1,5 +1,6 @@
 import $ from 'jquery'
 import Craft from 'craft'
+import Garnish from 'garnish'
 import Item from './Item'
 import NS from '../namespace'
 import BlockTypeFieldLayout from './BlockTypeFieldLayout'
@@ -14,6 +15,7 @@ export default Item.extend({
 
   _templateNs: [],
   _loaded: false,
+  _initialisedUi: false,
 
   init (settings = {}) {
     this.base(settings)
@@ -39,21 +41,11 @@ export default Item.extend({
     this.$moveButton = $neo.filter('[data-neo-bt="button.move"]')
     this.$actionsButton = $neo.filter('[data-neo-bt="button.actions"]')
 
-    // Set up the actions menu
-    this.$actionsButton.menubtn()
-    this._actionsMenu = this.$actionsButton.data('menubtn')
-    this._actionsMenu.on('optionSelect', e => this['@actionSelect'](e))
-    this.$actionsMenu = this._actionsMenu.menu.$container
-
-    // Stop the actions button click from selecting the block type and closing the menu
-    this.addListener(this.$actionsButton, 'click', e => e.stopPropagation())
-
-    if (settingsObj) {
-      settingsObj.on('change', () => this._updateTemplate())
-      settingsObj.on('destroy', () => this.trigger('destroy'))
-
-      this._updateTemplate()
-    }
+    // Manually select the block type, instead of letting the actions button click close the menu
+    this.addListener(this.$actionsButton, 'click', (e) => {
+      e.stopPropagation()
+      this.getField().selectItem(this)
+    })
 
     this.deselect()
   },
@@ -66,19 +58,20 @@ export default Item.extend({
     NS.leave()
     const errors = settings.getErrors()
     const hasErrors = (Array.isArray(errors) ? errors : Object.keys(errors)).length > 0
+    const menuId = `actionmenu${Math.floor(Math.random() * 1000000)}`
 
     return $(`
       <div class="nc_sidebar_list_item${hasErrors ? ' has-errors' : ''}" data-neo-bt="container.${this.getId()}">
         <div class="label" data-neo-bt="text.name">${settings.getName()}</div>
         <div class="smalltext light code" data-neo-bt="text.handle">${settings.getHandle()}</div>
         <a class="move icon" title="${Craft.t('neo', 'Reorder')}" role="button" data-neo-bt="button.move"></a>
-        <button class="settings icon menubtn" title="${Craft.t('neo', 'Actions')}" role="button" type="button" data-neo-bt="button.actions"></button>
-        <div class="menu" data-neo-bt="container.menu">
-          <ul class="padded">
-            <li><a data-icon="field" data-action="copy">${Craft.t('neo', 'Copy')}</a></li>
-            <li class="disabled"><a data-icon="brush" data-action="paste">${Craft.t('neo', 'Paste')}</a></li>
-            <li><a data-icon="share" data-action="clone">${Craft.t('neo', 'Clone')}</a></li>
-            <li><a class="error" data-icon="remove" data-action="delete">${Craft.t('neo', 'Delete')}</a></li>
+        <button class="btn action-btn menubtn" title="${Craft.t('neo', 'Actions')}" role="button" type="button" aria-controls="${menuId}" data-neo-bt="button.actions"></button>
+        <div id="${menuId}" class="menu menu--disclosure" data-neo-bt="container.menu">
+          <ul>
+            <li><button class="menu-item" data-icon="field" data-action="copy">${Craft.t('neo', 'Copy')}</button></li>
+            <li class="disabled"><button class="menu-item" data-icon="brush" data-action="paste">${Craft.t('neo', 'Paste')}</button></li>
+            <li><button class="menu-item" data-icon="share" data-action="clone">${Craft.t('neo', 'Clone')}</button></li>
+            <li><button class="menu-item error" data-icon="remove" data-action="delete">${Craft.t('neo', 'Delete')}</button></li>
           </ul>
         </div>
         <input type="hidden" name="${sortOrderName}[]" value="blocktype:${this.getId()}" data-neo-gs="input.sortOrder">
@@ -97,20 +90,26 @@ export default Item.extend({
    * @inheritDoc
    */
   load () {
+    if (`${this.getId()}`.startsWith('new')) {
+      // New block types should already be loaded
+      this._loaded = true
+    }
+
     if (this._loaded) {
-      // Already loaded
+      // Already loaded (though make sure the UI is initialised as well)
+      this._initUi()
       return Promise.resolve()
     }
 
     this.trigger('beforeLoad')
     const settings = this.getSettings()
     // Don't overwrite the field layout if it's already set (e.g. if pasting a block type)
-    const layout = this.getFieldLayout()?.getConfig() ?? settings.getFieldLayoutConfig()
-    const layoutId = settings.getFieldLayoutId()
+    const fieldLayout = this.getFieldLayout()?.getConfig() ?? settings.getFieldLayoutConfig()
+    const fieldLayoutId = settings.getFieldLayoutId()
     const data = {
       blockTypeId: this.getId(),
       errors: settings.getErrors(),
-      layout
+      fieldLayout
     }
 
     return new Promise((resolve, reject) => {
@@ -126,10 +125,10 @@ export default Item.extend({
 
           this._fieldLayout = new BlockTypeFieldLayout({
             namespace: [...this._templateNs, this._id],
-            html: response.data.layoutHtml,
-            id: layoutId,
+            html: response.data.fieldLayoutHtml,
+            id: fieldLayoutId,
             blockTypeId: data.blockTypeId,
-            initJs: false
+            initUi: false
           })
           this._settings.createContainer({
             html: response.data.settingsHtml.replace(/__NEOBLOCKTYPE_ID__/g, data.blockTypeId),
@@ -137,11 +136,76 @@ export default Item.extend({
           })
           this._loaded = true
 
+          if (this.getId()) {
+            this.getField().addItem(this)
+            this._initUi()
+          }
+
           this.trigger('afterLoad')
           resolve()
         })
         .catch(reject)
     })
+  },
+
+  _initUi () {
+    if (!this._loaded || this._initialisedUi) {
+      return
+    }
+
+    const settingsObj = this.getSettings()
+
+    // Set up the actions menu
+    this._actionsMenu = this.$actionsButton.data('trigger') || new Garnish.DisclosureMenu(this.$actionsButton)
+    this.$actionsMenu = this._actionsMenu.$container
+    this.addListener(this.$actionsMenu.find('[data-action]'), 'click', this['@actionSelect'])
+
+    if (settingsObj) {
+      settingsObj.on('change', () => this._updateTemplate())
+      settingsObj.on('destroy', () => this.trigger('destroy'))
+
+      this._updateTemplate()
+    }
+
+    // Make sure menu states (for pasting block types) are updated when changing browser tabs
+    const refreshPasteOptions = () => this.$actionsMenu
+      .find('[data-action="paste"]')
+      .parent()
+      .toggleClass('disabled', !window.localStorage.getItem('neo:copyBlockType'))
+    refreshPasteOptions()
+    this.addListener(document, `visibilitychange.blocktype${this.getId()}`, refreshPasteOptions)
+
+    this.getSettings()?.initUi()
+    this.getFieldLayout()?.initUi()
+    this._initialisedUi = true
+  },
+
+  /**
+   * @since 5.0.0
+   */
+  getConfig () {
+    const settings = this.getSettings()
+
+    return {
+      settings: {
+        childBlocks: settings.getChildBlocks(),
+        conditions: settings.getConditions(),
+        description: settings.getDescription(),
+        enabled: settings.getEnabled(),
+        iconId: settings.getIconId(),
+        ignorePermissions: settings.getIgnorePermissions(),
+        handle: settings.getHandle(),
+        minBlocks: settings.getMinBlocks(),
+        maxBlocks: settings.getMaxBlocks(),
+        minChildBlocks: settings.getMinChildBlocks(),
+        maxChildBlocks: settings.getMaxChildBlocks(),
+        minSiblingBlocks: settings.getMinSiblingBlocks(),
+        maxSiblingBlocks: settings.getMaxSiblingBlocks(),
+        name: settings.getName(),
+        topLevel: settings.getTopLevel()
+      },
+      fieldLayout: this.getFieldLayout().getConfig()
+    }
   },
 
   toggleSelect: function (select) {
@@ -157,7 +221,9 @@ export default Item.extend({
 
     if (fieldLayout) {
       fieldLayout.$container.toggleClass('hidden', !selected)
-    } else if (selected) {
+    }
+
+    if (selected) {
       this.load()
     }
 
@@ -175,15 +241,15 @@ export default Item.extend({
   },
 
   '@actionSelect' (e) {
-    const $option = $(e.option)
+    const option = e.target
 
-    if ($option.hasClass('disabled')) {
+    if (option.classList.contains('disabled')) {
       return
     }
 
-    this._actionsMenu?.hideMenu()
+    this._actionsMenu?.hide()
 
-    switch ($option.attr('data-action')) {
+    switch (option.getAttribute('data-action')) {
       case 'copy':
         this.trigger('copy')
         break